requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
django4-background-tasks==1.2.9
djangorestframework==3.14.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
